// session-web/app/utils/sessions.enum.ts
export enum ProgramStatus {
    OFFERING = 'offering',
    RUNNING = 'running',
    OFFBOARDING = 'offboarding',
  }
  
  export enum ProgramShortTitle {
    VC = 'vc',
    DATA = 'data',
    DATA2 = 'data2',
    DATA3 = 'data3',
    SCRUM = 'scrum',
    PRODUCT = 'product',
    PRODUCT2 = 'product2',
    GROWTH = 'growth',
  }
  